+++ hyperdbg/hprdbghv/code/memory/MemoryManager.c
@@ -107,12 +107,12 @@ MemoryManagerReadProcessMemoryNormal(HANDLE                    PID,
                 //
                 // Check whether the physical memory is valid or not
                 //
-                if (!CheckAddressPhysical(Address))
+                if (!CheckAddressPhysical((UINT64)Address))
                 {
                     return FALSE;
                 }
 
-                CopyAddress.PhysicalAddress.QuadPart = Address;
+                CopyAddress.PhysicalAddress.QuadPart = (LONGLONG)Address;
                 MmCopyMemory(UserBuffer, CopyAddress, Size, MM_COPY_MEMORY_PHYSICAL, ReturnSize);
             }
             else
